+++ locales/en-US.js
@@ -82,6 +82,9 @@ module.exports = {
 
   manifest: {
     not_found: "no such '%s' in current project",
+    circular_depends: "Circular dependency between %(system1)s and %(system2)s",
+    image_required: "Not image set for the `%(system)s' system",
+    depends_not_declared: "The `%(system)s` system depends on the `%(depend)s` system, which was not stated.",
   },
 
   system: {