+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   * [Docker] Add command `azk docker`
   * [Vm] Adding ssh keys generator, and uploads the key to start vm.
   * [Generators] Now Generators and rules is a extension of the UI.
+  * [Manifest] Validate: image is required.
+  * [Manifest] Validate: declared dependencies are required.
+  * [Manifest] Validate: circular dependencies are checked.
 
 * Bug
   * [Cli] Fix generate manifest in blank project dir.